requirements.txt
torch
torchvision
opencv-python
Pillow
matplotlib
tqdm